feat(orchestrator): Add agent instance IDs to GenesisResponse
This commit introduces a new field `agent_instance_ids` to the `GenesisResponse` model. This field is a dictionary that stores the process IDs of the agents that are run during the orchestration process.

The `orchestrator.py` file has been updated to collect the PIDs of the agent processes and include them in the `GenesisResponse` object.